requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
SQLAlchemy==1.3.22
Werkzeug==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
gunicorn